pyproject.toml
[tool.poetry]
name = "vaultctl"
version = "0.1.0"
description = ""
authors = ["Fernando Antunes <f.antunes84@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hvac = "^1.0.2"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
